https://github.com/jenkinsci/maven-hpi-plugin/pull/322#discussion_r839562552
