+++ src/main/java/org/jenkinsci/maven/plugins/hpi/InitializeMojo.java
@@ -1,11 +1,19 @@
 package org.jenkinsci.maven.plugins.hpi;
 
+import edu.umd.cs.findbugs.annotations.CheckForNull;
 import io.jenkins.lib.versionnumber.JavaSpecificationVersion;
+import java.io.File;
+import java.io.IOException;
 import java.util.ArrayList;
 import java.util.List;
+import java.util.jar.JarFile;
+import java.util.jar.Manifest;
+import org.apache.maven.artifact.Artifact;
 import org.apache.maven.plugin.MojoExecutionException;
 import org.apache.maven.plugins.annotations.LifecyclePhase;
 import org.apache.maven.plugins.annotations.Mojo;
+import org.apache.maven.shared.transfer.artifact.DefaultArtifactCoordinate;
+import org.apache.maven.shared.transfer.artifact.resolve.ArtifactResolverException;
 
 /**
  * Configure Maven for the desired version of Java.
@@ -51,5 +59,36 @@ public class InitializeMojo extends AbstractJenkinsMojo {
             }
         }
         return String.join(" ", arguments);
+    }
+
+    @CheckForNull
+    private String getAddOpens() throws MojoExecutionException {
+        Artifact artifact = resolveJenkinsWar();
+        File war = wrap(artifact).getFile();
+        try (JarFile jarFile = new JarFile(war)) {
+            Manifest manifest = jarFile.getManifest();
+            if (manifest == null) {
+                throw new MojoExecutionException("No manifest found in " + war);
+            }
+            return manifest.getMainAttributes().getValue("Add-Opens");
+        } catch (IOException e) {
+            throw new MojoExecutionException("Failed to read MANIFEST.MF from " + war, e);
+        }
+    }
+
+    private Artifact resolveJenkinsWar() throws MojoExecutionException {
+        DefaultArtifactCoordinate artifactCoordinate = new DefaultArtifactCoordinate();
+        artifactCoordinate.setGroupId("org.jenkins-ci.main");
+        artifactCoordinate.setArtifactId("jenkins-war");
+        artifactCoordinate.setVersion(findJenkinsVersion());
+        artifactCoordinate.setExtension("war");
+
+        try {
+            return artifactResolver
+                    .resolveArtifact(session.getProjectBuildingRequest(), artifactCoordinate)
+                    .getArtifact();
+        } catch (ArtifactResolverException e) {
+            throw new MojoExecutionException("Couldn't download artifact: ", e);
+        }
     }
 }
